feat(user): implement user management functionality
- Add useUser hook for user-related operations
- Implement updateUser, createGuest, softDeleteUser, and hardDeleteUser methods
- Integrate user state management with auth$
